requirements.txt
wheel
flask
psycopg2-binary
marshmallow
python-abc
python-dotenv
pytest~=7.2.0
pytest-cov~=4.0.0
pytest-mock~=3.10.0
pytest-env~=0.8.1
sqlalchemy
requests
importlib-metadata
flasgger
newrelic
# faker
# locust